+++ jeecg-boot-module-system/src/main/java/org/jeecg/modules/system/controller/SysTenantController.java
@@ -134,7 +134,7 @@ public class SysTenantController {
             // 过滤掉已被引用的租户
             List<String> idList = new ArrayList<>();
             for (String id : ls) {
-                int userCount = sysTenantService.countUserLinkTenant(id);
+                long userCount = sysTenantService.countUserLinkTenant(id);
                 if (userCount == 0) {
                     idList.add(id);
                 }
