+++ jeecg-boot-module-system/src/main/java/org/jeecg/modules/system/service/ISysTenantService.java
@@ -22,7 +22,7 @@ public interface ISysTenantService extends IService<SysTenant> {
      * @param id
      * @return
      */
-    int countUserLinkTenant(String id);
+    long countUserLinkTenant(String id);
 
     /**
      * 根据ID删除租户，会判断是否已被引用
