+++ jeecg-boot-module-system/src/main/java/org/jeecg/modules/system/service/impl/SysPermissionServiceImpl.java
@@ -42,7 +42,7 @@ public class SysPermissionServiceImpl extends ServiceImpl<SysPermissionMapper, S
 
 	@Resource
 	private SysPermissionMapper sysPermissionMapper;
-	
+
 	@Resource
 	private ISysPermissionDataRuleService permissionDataRuleService;
 
@@ -73,7 +73,7 @@ public class SysPermissionServiceImpl extends ServiceImpl<SysPermissionMapper, S
 		}
 		String pid = sysPermission.getParentId();
 		if(oConvertUtils.isNotEmpty(pid)) {
-			int count = this.count(new QueryWrapper<SysPermission>().lambda().eq(SysPermission::getParentId, pid));
+			long count = this.count(new QueryWrapper<SysPermission>().lambda().eq(SysPermission::getParentId, pid));
 			if(count==1) {
 				//若父节点无其他子节点，则该父节点是叶子节点
 				this.sysPermissionMapper.setMenuLeaf(pid, 1);
@@ -94,10 +94,10 @@ public class SysPermissionServiceImpl extends ServiceImpl<SysPermissionMapper, S
 		//删除部门角色授权
 		sysDepartRolePermissionMapper.deleteByMap(map);
 	}
-	
+
 	/**
 	 * 根据父id删除其关联的子节点数据
-	 * 
+	 *
 	 * @return
 	 */
 	public void removeChildrenBy(String parentId) {
@@ -108,7 +108,7 @@ public class SysPermissionServiceImpl extends ServiceImpl<SysPermissionMapper, S
 		List<SysPermission> permissionList = this.list(query);
 		if (permissionList != null && permissionList.size() > 0) {
 			String id = ""; // id
-			int num = 0; // 查出的子级数量
+			long num = 0; // 查出的子级数量
 			// 如果查出的集合不为空, 则先删除所有
 			this.remove(query);
 			// 再遍历刚才查出的集合, 根据每个对象,查找其是否仍有子级
@@ -132,7 +132,7 @@ public class SysPermissionServiceImpl extends ServiceImpl<SysPermissionMapper, S
 			}
 		}
 	}
-	
+
 	/**
 	  * 逻辑删除
 	 */
@@ -145,7 +145,7 @@ public class SysPermissionServiceImpl extends ServiceImpl<SysPermissionMapper, S
 			throw new JeecgBootException("未找到菜单信息");
 		}
 		String pid = sysPermission.getParentId();
-		int count = this.count(new QueryWrapper<SysPermission>().lambda().eq(SysPermission::getParentId, pid));
+		long count = this.count(new QueryWrapper<SysPermission>().lambda().eq(SysPermission::getParentId, pid));
 		if(count==1) {
 			//若父节点无其他子节点，则该父节点是叶子节点
 			this.sysPermissionMapper.setMenuLeaf(pid, 1);
@@ -189,29 +189,29 @@ public class SysPermissionServiceImpl extends ServiceImpl<SysPermissionMapper, S
 				sysPermission.setParentId("");
 			}
 			//Step2.判断菜单下级是否有菜单，无则设置为叶子节点
-			int count = this.count(new QueryWrapper<SysPermission>().lambda().eq(SysPermission::getParentId, sysPermission.getId()));
+			long count = this.count(new QueryWrapper<SysPermission>().lambda().eq(SysPermission::getParentId, sysPermission.getId()));
 			if(count==0) {
 				sysPermission.setLeaf(true);
 			}
 			//----------------------------------------------------------------------
 			this.updateById(sysPermission);
-			
+
 			//如果当前菜单的父菜单变了，则需要修改新父菜单和老父菜单的，叶子节点状态
 			String pid = sysPermission.getParentId();
 			if((oConvertUtils.isNotEmpty(pid) && !pid.equals(p.getParentId())) || oConvertUtils.isEmpty(pid)&&oConvertUtils.isNotEmpty(p.getParentId())) {
 				//a.设置新的父菜单不为叶子节点
 				this.sysPermissionMapper.setMenuLeaf(pid, 0);
 				//b.判断老的菜单下是否还有其他子菜单，没有的话则设置为叶子节点
-				int cc = this.count(new QueryWrapper<SysPermission>().lambda().eq(SysPermission::getParentId, p.getParentId()));
+				long cc = this.count(new QueryWrapper<SysPermission>().lambda().eq(SysPermission::getParentId, p.getParentId()));
 				if(cc==0) {
 					if(oConvertUtils.isNotEmpty(p.getParentId())) {
 						this.sysPermissionMapper.setMenuLeaf(p.getParentId(), 1);
 					}
 				}
-				
+
 			}
 		}
-		
+
 	}
 
 	@Override
@@ -226,9 +226,9 @@ public class SysPermissionServiceImpl extends ServiceImpl<SysPermissionMapper, S
 	public void deletePermRuleByPermId(String id) {
 		LambdaQueryWrapper<SysPermissionDataRule> query = new LambdaQueryWrapper<>();
 		query.eq(SysPermissionDataRule::getPermissionId, id);
-		int countValue = this.permissionDataRuleService.count(query);
+		long countValue = this.permissionDataRuleService.count(query);
 		if(countValue > 0) {
-			this.permissionDataRuleService.remove(query);	
+			this.permissionDataRuleService.remove(query);
 		}
 	}
